feat(wip): add allocation amount input and percentage
