+++ src/app/collective-rewards/allocations/Header.tsx
@@ -1,0 +1,13 @@
+import { HeaderTitle, Label, Paragraph, Typography } from '@/components/Typography'
+
+export const Header = () => {
+  return (
+    <>
+      <HeaderTitle>Confirm strif allocation for the selected builders</HeaderTitle>
+      <Paragraph className="font-normal leading-6 text-[rgba(255,255,255,0.6)]">
+        Back builders who drive innovations. Review the selection and support the Builders you align with.
+        Your allocations influence the rewards the Builders will receive.
+      </Paragraph>
+    </>
+  )
+}
